improvement: create a route to return genres and upgrade search
We sort propositions descending by peers and seeds sum and rating in
descending mode too.

We implemented the searching mode.
The following criteria can be used :
- `query`: a string query matching movies titles
- `genre`: a string allowing to get all movies tagged as being of this
genre
- `year`: a number allowing to get all movies released this year

During the seeding we compute `peersAndSeedsCount` property, normalize
the genres and filter unkwown genres. Thanks to this filtering movies
for adults are removed.