feat(movie): implement the movie player
The user can select which torrent to load.
After that, film downloading starts and when sufficient bytes have been
saved on the server, the browser request them.
To be able to download the tracks we had to put the attribute
`crossorigin` of `video` tag to `use-credentials`. But now the poster
can't be downloaded for security reasons as its saved on a server
responding with insecure CORS headers (IMDB).